add testdata

+++ python/google_OCR.py
@@ -37,10 +37,11 @@ def main(args):
     args.out_dir.mkdir(exist_ok=True, parents=True)
     all_img_paths = args.img_dir.glob("*.png")
     for img_path in all_img_paths:
+        out_path = args.out_dir / (img_path.stem + ".txt")
+        if out_path.exists():
+            continue
         print("Processing", str(img_path))
         texts = detect_text(img_path)
-        out_path = args.out_dir / (img_path.stem + ".txt")
-
         with open(out_path, "w", encoding="utf8") as f:
             for text in texts:
                 t = text.description.encode('cp932', "ignore")
